+++ ABJC/UI/Preferences/Preferences+Client.swift
@@ -21,6 +21,15 @@ extension PreferencesView {
                 Section(header: Label("pref.client.ui.label", systemImage: "uiwindow.split.2x1")) {
                     InfoToggle("pref.client.showstitles.label", "pref.client.showstitles.descr", $session.preferences.showsTitles)
                     
+//                    InfoToggle("pref.client.backdroptitleimages.label", "pref.client.backdroptitleimages.descr", $session.preferences.posterT)
+                    
+                    Picker("pref.client.postertype.label", selection: $session.preferences.posterType) {
+                        ForEach(PreferenceStore.PosterType.allCases, id: \.rawValue) { value in
+                            Text(value.localizedName)
+                                .tag(value)
+                        }
+                    }
+                    
                     Picker("pref.client.collectiongrouping.label", selection: $session.preferences.collectionGrouping) {
                         ForEach(Grouping.allCases, id: \.rawValue) { value in
                             Text(value.localizedName)