+++ ABJC/UI/Subviews/Episode Row & Card/EpisodeCardView.swift
@@ -16,12 +16,13 @@ extension LibraryView
         
         /// Size
         private var size: CGSize = CGSize(width: 548, height: 308.25 )
+//        private var size: CGSize = CGSize(width: 640, height: 360 )
         
         /// Media Item
         private var item: APIModels.Episode
         
         private var url: URL {
-            return API.imageURL(session.jellyfin!, item.id, .thumb)
+            return API.imageURL(session.jellyfin!, item.id, .primary)
         }
         /// Initializer
         /// - Parameter item: Item