+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-build.22] - [2021-05-20]
+### Added
+- It is now possible to change the type of poster that is displayed for media items
+- Item detail pages have a backdrop (Blurhash)
+- Item detail pages show a poster for the item
+### Changed
+- Updated URLImage to Version 2.2.5
+### Deprecated
+### Fixed 
+### Removed
+
 ## [v1.0.0-build.22] - [2021-05-10]
 ### Added
 ### Changed